Produce a confusion matrix when verbose is True
I discovered that the model was accurate because it always guessed
that the patient did not die.  That's good for the r-squared score,
but terrible in practice.  I used an under sampling technique to try
and prevent it from doing so and now it performs very poorly.

I think that this is an artefact of using the test data set more than
anything else.